requirements.txt
ruamel.yaml
addict
Click>=8.0
requests
jsonschema>=3.0.0
flask
sh
rich
rich-click
pygit2; python_version >= '3.7'
